refactor: Moves theme logic to runtime file
