+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import '@/styles/globals.css'
 import { Background } from '@/components/layout/background'
 import { Footer } from '@/components/layout/footer'
 import { Header } from '@/components/layout/header'
-import { buildInitThemeScript, getDefaultThemeForNow } from '@/lib/themes'
+import { buildInitThemeScript, getDefaultThemeForNow } from '@/lib/theme-runtime'
 import { type Metadata } from 'next'
 import { Noto_Sans, Space_Grotesk } from 'next/font/google'
 import { cookies } from 'next/headers'
